Source code for the Shift Calendar
This should be a good start for the hackathon.  I tried to sanitize
it by removing anything that is unlikely to be useful, or which
could compromise Shift's data.  The account information is all
located in include/account.php, which contains info from my own
system, not shift2bikes.org, so it should be a safe example.